fix clang -Wconversion
src/dbcache.c:38:12: error: implicit conversion loses integer precision: 'int' to 'short' [-Werror,-Wconversion]
        n->sync = sync;

+++ src/dbcache.c
@@ -5,7 +5,7 @@
 #include "cfg.h"
 #include "ibw.h"
 
-int cacheadd(const char *iface, int sync)
+int cacheadd(const char *iface, short sync)
 {
 	datanode *p, *n;
